V2.8.20 Refactor
